# FastRegexMatcher: Extract longer prefixes